requirements.txt
pause
python-vlc
ntplib